requirements.txt
confluent-kafka==2.3.0
pydantic==2.5.2
pytest==7.4.3
pytest-cov==4.1.0
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-pubsub
google-cloud-storage
google-cloud-bigquery
opentelemetry-api
opentelemetry-sdk